Documentation: panels nowadays can display more than 2 columns.
In the past panels supported up to 2 "columns" only. We update the
documentation.

Note: a naive fix is to change it to say "After the panel size, you may specify
the number of columns to display in the panel". Unfortunately, users may think
by "columns" we mean "fields", so we have to word it differrently.

Signed-off-by: Mooffie <mooffie@gmail.com>