Ticket # 4060: escape backslashes in PHP single quote string.
In PHP strings delimited with single quotes, there are two characters
that can be escaped: \ and '. For example, the PHP string 'a\'b\\c\d'
could be printed as ‘a'b\c\d’.

Escaping quotes was already possible with the PHP syntax file. This
commit adds support for escaping backslashes. It fixes mcedit’s syntax
highlighting on PHP strings that end with \\ (like 'aaa\\').

Signed-off-by: Andrew Borodin <aborodin@vmail.ru>