Ticket #4087: verilog syntax: removing keyboards attribute, endattribute, strength
Keywords attribute, endattribute, strength are not a reserved keywords in any
verilog standard. I can find a few references to these saying that they
were reserved keywords in OVI LRM version 2.0 (a version of verilog
before it was standardized), but this LRM is not available. The OVI LRM
version 1.0 didn't include these keywords, and none of the IEEE standards
include these keywords. I also tried the syntax highlighter of two
commercial tools, and they don't show attribute/endattribute as keywords.

Signed-off-by: Andrew Borodin <aborodin@vmail.ru>