fix endianess bug in handle_console_linux()
handle_console_linux() passes the address of action to read() and intends
to send one byte of data (the LSB). On LE this works, but one BE machines
this sends the MSB which is always zero. Fix this by assigning the value
first to a char type, and use that instead.

Signed-off-by: Sven Schnelle <svens@stackframe.org>