Ticket #4626: mc.ext.ini: support OpenEmbedded ipk archives
Two types of ipk archives appear to exist:
* OpenWRT
* Openembedded / Yocto

The former is a (gzip-compressed) tar archive, which appears to be
modelled after the Debian package format (but not matching it
exactly!), containing the three files debian-binary, data.tar, and
control.tar (the latter two may be compressed).

The latter is a proper Debian archive, which means it's an ar archive,
with the files debian-binary, control.tar, and data.tar in exactly this
order. control and data may also be compressed.

While modelled after the Debian package format [1], OpenWRT ipk
archives don't strictly adhere to it: Debian packages should be ar
archives (not tar) and the order of files inside the archive matters
(OpenWRT doesn't follow this).

Midnight Commander is currently unable to display the Openembedded ipk
archives as both use the .ipk suffix, and because one is a tar- and one
is an ar archive, and because the ipk suffix is hooked up to
Include=tar.gz, which (obviously) doesn't support ar archives.

Update mc.ext.ini to also handle OpenEmbedded ipk packages by also
inspecting the type, in addition to the suffix. Note that we can not
unconditionally use the deb rule for both types because the OpenWRT
package don't strictly conform to the Debian package format and
dpkg-deb and friends complain.

Link: https://manpages.debian.org/bookworm/dpkg-dev/deb.5.en.html [1]
Signed-off-by: André Draszik <andre.draszik@linaro.org>
Signed-off-by: Andrew Borodin <aborodin@vmail.ru>