Ticket #4627: fix segfault in the help engine
...when going to the previous topic.

Inside help_interactive_display(), the for loop that initializes history
leaves history_ptr at -1.

In help_back(), this value causes an out-of-bounds access to the history
array, thus placing random noise in currentpoint and selected_item, the
former used subsequently by help_bg_callback() and passed to
help_show().

  * (history_ptr): initialize to 0.
  * (help_interactive_display): use separate variable for loop that
  initializes history.

Thanks Egmont Koblinger for the finding out the cause of segfault.

Signed-off-by: Andrew Borodin <aborodin@vmail.ru>