Ticket #4632: fix escape sequences for Shift-Fx keys on Linux console
Note that due to an old tradition spanning over many decades, in mc's speak we use F11 for Shift-F1, F12 for Shift-F2 and so on. Physical F11/F12 keys cannot be currently mapped, even if they do exist on the keyboard and produce discernible escape sequences.

This commit fixes an off-by-two situations with Shift-Fx keys on Linux consoles. All modern consoles in the distributions we know of now produce these sequences.

Signed-off-by: Yury V. Zaytsev <yury@shurup.com>