buildsys: add some patterns to `.gitignore`
There are some build output files that are not gitignore; most are generated
by "make", except for *.trs which is from "make -C tests/src/editor/ check",
and "conftest.c" and "confdefs.h" which is from a configure step.

The missing ignores makes it very easy to accidentally commit files
when working with the Jujutsu VCS which tracks files automatically.
Fix this ignoring all patterns I've seen.

Signed-off-by: Johannes Altmanninger <aclopte@gmail.com>
Signed-off-by: Yury V. Zaytsev <yury@shurup.com>