Extract escape-sequence parsing logic into new module
Following commits want to add more logic (and tests) for decoding terminal
escape sequences, and implement a subset of the kitty keyboard protocol.
It's probably not right to keep all of this in lib/util.c, so extract a
new module.

I called it "terminal" because it helps mc act as terminal.

We also have lib/tty/tty.h though that's for talking to the actual terminal
that mc is running inside.

Signed-off-by: Johannes Altmanninger <aclopte@gmail.com>