subshell: tell fish that we actually support the kitty keyboard protocol now
__mc_csi_u was added so that fish shell knows when it's safe to enable the
kitty keyboard protocol.

As seen in the parent commit, it wasn't quite safe; so the upcoming patch
release of fish will disable this protocol again -- until fish finds this
commit, so add a new version marker.

While at it, use a less misleading name.  Dropping the old name will mean that
fish < 4.0.1 will no longer enable the kitty keyboard protocol (and be safe
from the hang on numlock). I guess we could avoid that but I don't think this
is not a big loss, especially because since fish 4.0.1 will enable it again.

Signed-off-by: Johannes Altmanninger <aclopte@gmail.com>