subshell: reorder pty/pipe file descriptor acquisition
Swap the code allocating the pty and the block getting
the pipe for cwd communication, to hopefully get an fd
lower than 10 for the latter, to ameliorate a few from
the problems mentioned in #4634 as long as there isn’t
more global code allocating file descriptors added.

Specifically, this moves the shell’s side of the fd in
my local tests on Debian trixie from 10 to 8.

I’ve merely swapped two blocks without changing any of
the actual code. Swapping WRITE with READ in line 158ﬀ
would gain us another one in this specific instance as
it would then use the lower number of the pair.

https://github.com/MidnightCommander/mc/issues/4634#issuecomment-2989003485
mentions a few ideas I have for really actually fixing
this in the long term (and ensuring the issue will not
reappear) which need more invasive refactoring though.

Signed-off-by: Thorsten Glaser <tglaser@b1-systems.de>