Ticket #4726: add mc-wrapper for fish shell
fish shell occasionally gets bug reports from mc users, so there are
definitely some mc+fish users, but none of them has noticed the lack of
"/usr/libexec/mc/mc-wrapper.fish".

That file is supposed to define an alias that behaves like mc except that,
after mc exits, it will change fish's directory to the one last used by mc.

Users need to source this manually; if this is the behavior that most people
expect, we should make it the default (for example by shipping this function
with fish).

For now, add it to mc, which shouldn't break anyone even if fish decides to
add this as well.

Note that fish does not do word splitting, so variable expansions usually
don't need to be quoted.

Tested as follows:

	$ make install
	$ type mc
	mc is /home/johannes/.local/bin/mc
	$ source /home/johannes/.local/libexec/mc/mc.fish
	$ type mc
	mc is a function with definition
	# Defined in /home/johannes/.local/libexec/mc/mc.fish @ line 1
	function mc --description 'Visual shell for Unix-like systems - fish wrapper'
		source /home/johannes/.local/libexec/mc/mc-wrapper.fish $argv
	end
	mc
	# navigate somewhere, press F10 to exit

Co-authored-by: Jarkko Sakkinen <jarkko.sakkinen@iki.fi>

Closes #4726.

Co-authored-by: Jarkko Sakkinen <jarkko.sakkinen@iki.fi>
Signed-off-by: Johannes Altmanninger <aclopte@gmail.com>
Signed-off-by: Yury V. Zaytsev <yury@shurup.com>