contrib: wrappers to propagate mc exit code (#1466)
As reported in [1] if I paste this command into an interactive Bash shell:

	( sleep 1 && pkill mc ) &
	. ~/.local/libexec/mc/mc.sh
	mc; s=$?; reset; echo exit-status=$s

I get "exit-status=0" which is wrong.

Also, if "set -e" is active (which is unlikely), we fail to cd or clean up
temporary files.

Fix both by capturing the error code from mc, and propagating it only at
the very end of the wrapper.
Take care to not leak the MC_STATUS variable, to not pollute "set" output.

Tested on bash, zsh and tcsh.

[1]: https://github.com/MidnightCommander/mc/pull/4726#discussion_r2176458948

Signed-off-by: Yury V. Zaytsev <yury@shurup.com>
Signed-off-by: Johannes Altmanninger <aclopte@gmail.com>