Ticket #4752: fix crash when loading corrupted file position history
Add NULL pointer check to `file_history_parse_entry` to prevent a crash when
lines are encountered that do not match expected format.

Further, add escaping and unescaping to `file_history_parse_entry` and
`file_history_write_entry` to handle the file names with special characters.

Likewise, add escaping to `load_file_position` and `save_file_position`. In
`load_file_position` unescaping is not needed, because file names are only
compared to the ones in the history file.

Unit tests are added for `file_history_*_entry` functions. Tests for
`*_file_position` functions are only possible with a major refactoring.

Signed-off-by: Yury V. Zaytsev <yury@shurup.com>