Ticket #3792, #4634: Do not override the subshell's prompt
Do not override 'PS1' (bash and zsh) and 'prompt' (tcsh). This is unkind
behavior (we should respect the user's preference), and due to a bug that
is fixed in the next commit (ticket #4625) it works only for tcsh anyway.

Additional comments by Yury regarding tcsh:

This override was sneaked in #2742 without any explanation, presumably for
consistency and/or due to confusion with our non-working fallbacks for other
shells.

In as far as fallbacks are concerned, I still haven't gotten to the bottom of
it as to why they were introduced in the first place. By reading some old code,
my impression was that the idea was to provide consistent experience across
shells. Then probably the code was misinterpreted to mean a fallback if no
prompt is defined. And finally other changes broke it anyways and it was
carried further as a cargo cult. Even later overrides were removed one by one
after the users complained, and finally it seems there is only tcsh one left.
So much for a vague backstory on this change.

Relevant commit: f596c916a42a0868897b3314e557b0a82df37017

Signed-off-by: Yury V. Zaytsev <yury@shurup.com>
Signed-off-by: Egmont Koblinger <egmont@gmail.com>