lib/tty/key.c: Add Kitty Keyboard Protocol support
The protocol is described here:

https://sw.kovidgoyal.net/kitty/keyboard-protocol/

The old functionality that "eats" all characters following an unknown
escape sequence has been removed, as it causes severe side effects (such
as not responding to any keys following an unknown sequence for some
time). This is superseded by parsing CSI and SS3 sequences, which are
the de facto standard for sending any function keys.

Signed-off-by: Jakub Horký <jakub.git@horky.net>