Fallout from PR #4784: Force a subshell cd command at startup
Previously we omitted injecting a cd command to the subshell at startup
if it was already in the target directory. The first prompt of the subshell
was read and displayed in mc's panels.

This plays badly with the forthcomming change where testing the persistent
buffer code will read and discard the subshell's data to avoid a deadlock.

The proper solution would be more complicated than justified. Just get a
new prompt printed and we'll catch that.

Signed-off-by: Egmont Koblinger <egmont@gmail.com>