Ticket #3158, #4799: Clean up and fix frame drawing
Fix the ncurses build to show double frame characters in UTF-8 mode
wherever the skin requests that.

Fall back to single frame characters, rather than displaying random
letters, if the value specified in the skin file cannot be converted to
the current locale.

Internal refactoring and cleanup: Use constants that are independent from
the screen drawing library. Separate the roles (e.g. "the character to be
placed in the upper left corner of a supposedly doubly-framed dialog")
from the actual characters (e.g. "double upper left corner line").

Minor regression due to the refactoring: Double lines are replaced by
single ones in slang build under KOI8-R locale. It's probably not worth it
to fix this.

Signed-off-by: Egmont Koblinger <egmont@gmail.com>