Ticket #4799: Fix special skin characters in 8-bit locales
Most skins define some fancy Unicode characters for places like to denote
if a filename doesn't fit, or for the scrollbar's arrow, etc.

In 8-bit locales these appeared as two or three garbage characters. Fix it
by falling back to the default ASCII value wherever the skin's character
can't be represented in the current 8-bit locale.

Signed-off-by: Egmont Koblinger <egmont@gmail.com>