lib/tty: Add true color support with ncurses and improve S-Lang color detection
Enable true color support in ncurses. TERM must be set to a relevant
terminfo entry, i.e. *-direct, *-direct16, or *-direct256.

When set to *-direct256, both 256-color and true color mode can be used
at the same time.

Also fix true color detection in S-Lang. S-Lang checks for RGB terminfo
capability to enable true color, so we should check for it as well. This
means the COLORTERM variable doesn't need to be set when using direct
terminfo variants. This also applies to simultaneous 256-color and true
color mode.

We need to access the RGB extended capability. This is supported in both
S-Lang and ncurses, so unify the terminfo access routines to enable it.
Now they should be called with both terminfo and termcap capability
name.

Update FAQ.

Resolves: #4137
Resolves: #4821

Signed-off-by: Jakub Horký <jakub.git@horky.net>