lib/tty/key.c: Replace paranoia code with parsing code
The old functionality that "eats" all characters following an unknown
escape sequence is removed, as it causes severe side effects (such
as not responding to any keys following an unknown sequence for some
time). This is superseded by parsing CSI and SS3 sequences, which are
the de facto standard for sending any function keys.

Resolves: #3136

Signed-off-by: Jakub Horký <jakub.git@horky.net>